Patch race condition
Argh, I broke the netgroup and building lists! This code relies on delays everywhere because the way it's structured has race conditions! I took away a delay thinking it would be fine but it re-introduced a race condition and broke the data inside the Netgroups and Port Search cards. I'll try to make it not rely on hard delays some time.