Add phone number formatter
New card for the Help Desk: Format Number 📞. Enter a phone number and it will format it to the specific format that Azure looks for. This is helpful for when you get a number like "+2345678901234" and you can't tell where the country code ends and the rest of the phone number begins (this one is +234).