Add link to Country Codes wiki page
Clicking the Format Number tool's title will now take you to the Wikipedia page that I used to use to look up country codes. I thought it might be helpful just in case you want to look at the whole list, or if the tool stops working for whatever reason.
Also Format Number was the only card on the page whose title wasn't clickable so now it fits in better