Verify operation mode value
The old site uses numbers for this value, while the current version uses strings. If coming from the old site, the mode selector will get populated with a number and appear blank, when it should default to Help Desk. This fixes that by properly defaulting the value to Help Desk if it is anything other than one of the operation mode slugs.